# Conflicts:
#	src/main/java/aaa/microfinance/server/controller/LendDetailController.java
#	src/main/java/aaa/microfinance/server/controller/RerecordsController.java
#	src/main/java/aaa/microfinance/server/dao/LendDetailDao.java
#	src/main/java/aaa/microfinance/server/dao/RerecordsDao.java
#	src/main/java/aaa/microfinance/server/entity/Customer.java
#	src/main/java/aaa/microfinance/server/entity/Installment.java
#	src/main/java/aaa/microfinance/server/entity/LendDetail.java
#	src/main/java/aaa/microfinance/server/entity/Loanorders.java
#	src/main/java/aaa/microfinance/server/service/LendDetailService.java
#	src/main/java/aaa/microfinance/server/service/RerecordsService.java
#	src/main/java/aaa/microfinance/server/service/impl/LendDetailServiceImpl.java
#	src/main/java/aaa/microfinance/server/service/impl/RerecordsSerciveImpl.java
#	src/main/resources/mapper/LendDetailDaoImpl.xml
#	src/main/resources/mapper/RerecordsDaoImpl.xml
#	src/main/resources/templates/background/sysLendDetail.html

+++ src/main/java/aaa/microfinance/server/entity/Rerecords.java
@@ -13,20 +13,23 @@ public class Rerecords {
     private Long cusid;
     private String cusname;
     private Long loanmount;
+
     @JsonFormat(pattern="yyyy-MM-dd HH:mm")
     private Date orderdate;
+
     @JsonFormat(pattern="yyyy-MM-dd HH:mm")
     private Date reimdate;
-    private Long bystages;
+
+    private Integer bystages;
     private Integer loanstatus;
     private String ordernumber;
-    private Long totalpay;
+    private Double totalpay;
 
-    public Long getTotalpay() {
+    public Double getTotalpay() {
         return totalpay;
     }
 
-    public void setTotalpay(Long totalpay) {
+    public void setTotalpay(Double totalpay) {
         this.totalpay = totalpay;
     }
 
@@ -86,11 +89,11 @@ public class Rerecords {
         this.reimdate = reimdate;
     }
 
-    public Long getBystages() {
+    public Integer getBystages() {
         return bystages;
     }
 
-    public void setBystages(Long bystages) {
+    public void setBystages(Integer bystages) {
         this.bystages = bystages;
     }
